#10 Convert Logger to a DLL Library
