+++ EasySave/EasySave/EasySave.csproj
@@ -8,7 +8,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <ProjectReference Include="..\Logger\Logger.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="Logger">
+      <HintPath>..\Logger\bin\Debug\net8.0\Logger.dll</HintPath>
+    </Reference>
   </ItemGroup>
 
 </Project>
