+++ EasySave/EasySave/FullSave.cs
@@ -10,7 +10,7 @@ namespace EasySave
     {
         //ATTRIBUTES
 
-
+        
 
         //CONTRUCTOR
         public FullSave(string name, string sourcePath, string targetPath) : base(name, sourcePath, targetPath)
